requirements.txt
langchain
langchain-community
langchain-google-genai
pypdf
faiss-cpu
google-generativeai
python-dotenv